Title: 함수를 이용한 369 게임, Time: 132ms, Memory: 10MB, Status: Wrong Answer - Codetree

+++ 250714/함수를 이용한 369 게임/369-games-using-functions.java
@@ -25,7 +25,7 @@ public class Main {
         int B = sc.nextInt();
         // Please write your code here.
 
-        int count = 0;
+        long count = 0;
         for(int i = A; i <= B; i++){
             if(isMagicNum(i)) count++;
         }
